[#15] refactor: reorganize source tree into app/walker/render/sdk
