+++ src/render/HtmlRenderer.h
@@ -5,7 +5,7 @@
 #include <map>
 #include <iostream>
 
-#include "SDK_Wrapper.h"
+#include "sdk/SDK_Wrapper.h"
 
 namespace Html {
